move favourites context menus to favourites module
