+++ xbmc/ContextMenus.cpp
@@ -18,13 +18,9 @@
  *
  */
 
-#include "FileItem.h"
-#include "ServiceBroker.h"
-#include "dialogs/GUIDialogFavourites.h"
 #include "storage/MediaManager.h"
-#include "utils/URIUtils.h"
-
 #include "ContextMenus.h"
+
 
 namespace CONTEXTMENU
 {
@@ -55,43 +51,6 @@ namespace CONTEXTMENU
   bool CEjectDrive::Execute(const CFileItemPtr& item) const
   {
     return g_mediaManager.Eject(item->GetPath());
-  }
-
-  bool CFavouriteContextMenuAction::IsVisible(const CFileItem& item) const
-  {
-    return URIUtils::IsProtocol(item.GetPath(), "favourites");
-  }
-
-  bool CFavouriteContextMenuAction::Execute(const CFileItemPtr& item) const
-  {
-    CFileItemList items;
-    CServiceBroker::GetFavouritesService().GetAll(items);
-    for (const auto& favourite : items)
-    {
-      if (favourite->GetPath() == item->GetPath())
-      {
-        if (DoExecute(items, favourite))
-          return CServiceBroker::GetFavouritesService().Save(items);
-      }
-    }
-    return false;
-  }
-
-  bool CRemoveFavourite::DoExecute(CFileItemList &items, const CFileItemPtr& item) const
-  {
-    int iBefore = items.Size();
-    items.Remove(item.get());
-    return items.Size() == iBefore - 1;
-  }
-
-  bool CRenameFavourite::DoExecute(CFileItemList&, const CFileItemPtr& item) const
-  {
-    return CGUIDialogFavourites::ChooseAndSetNewName(item);
-  }
-
-  bool CChooseThumbnailForFavourite::DoExecute(CFileItemList&, const CFileItemPtr& item) const
-  {
-    return CGUIDialogFavourites::ChooseAndSetNewThumbnail(item);
   }
 
 } // namespace CONTEXTMENU
